refactor: RowMapper에서 도메인 객체를 반환할 수 있도록 변경
